requirements.txt
Pillow
imageio
pre-commit
pytest
argparse
numpy
pygifsicle